requirements.txt
django
djangorestframework
requestscd
dataclasses